% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDSF}
\alias{BDSF}
\title{Losses on Business Disruption and System Failures}
\format{
\code{numeric} with 152 operational losses.
}
\description{
Operational losses from the \emph{business disruption and system failures} of
the Italian bank UniCredit between 2005 and 2014. The data, which are scaled
by an unknown factor for anonymity reasons, have been provided by the
UniCredit's Operational Risk Department. More information related to these
data can be found in \insertCite{hambuckers2018;textual}{tukeyGH} and
\insertCite{bee2021a;textual}{tukeyGH}.
}
\references{
\insertAllCited{}
}
\seealso{
Other data: 
\code{\link{CPBP2014}},
\code{\link{EDPM2014}},
\code{\link{EPWS2014}}
}
\concept{data}
\keyword{data}
