\name{dgh}
\alias{dgh}
\title{Density function for Tukey's g-and-h
Currently working for h=0 only}
\usage{
  dgh(x, g = 0, h = 0, A = 0, B = 1)
}
\arguments{
  \item{x}{numeric}

  \item{g}{parameter}

  \item{h}{parameter}

  \item{A}{parameter}

  \item{B}{parameter}
}
\description{
  Density function for Tukey's g-and-h Currently working
  for h=0 only
}

