% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EDPM2014}
\alias{EDPM2014}
\title{Losses on Execution, Delivery, and Process Management}
\format{
\code{numeric} with 417 operational losses.
}
\description{
Operational losses from the \emph{execution, delivery, and process management} of
the Italian bank UniCredit in 2014. The data, which are scaled by an unknown
factor for anonymity reasons, have been provided by the UniCredit's
Operational Risk Department. Losses are adjusted for inflation
by means of the monthly Italian consumer price index (base 100: December
2013). Around one-fourth of the losses are related to financial instruments
and derivative products. More information related to these data can be found
in \insertCite{hambuckers2018;textual}{tukeyGH} and
\insertCite{bee2019a;textual}{tukeyGH}.
}
\references{
\insertAllCited{}
}
\seealso{
Other data: 
\code{\link{BDSF}},
\code{\link{CPBP2014}},
\code{\link{EPWS2014}}
}
\concept{data}
\keyword{data}
