% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPWS2014}
\alias{EPWS2014}
\title{Losses on Employment Practices and Workplace Safety}
\format{
\code{numeric} with 97 operational losses.
}
\description{
Operational losses from the \emph{employment practices and workplace safety} of
the Italian bank UniCredit in 2014. The data, which are scaled by an unknown
factor for anonymity reasons, have been provided by the UniCredit's
Operational Risk Department. More information related to these data can be
in \insertCite{hambuckers2018;textual}{tukeyGH} and
\insertCite{bee2019a;textual}{tukeyGH}.
}
\references{
\insertAllCited{}
}
\seealso{
Other data: 
\code{\link{BDSF}},
\code{\link{CPBP2014}},
\code{\link{EDPM2014}}
}
\concept{data}
\keyword{data}
