\name{goodnessFit}
\alias{goodnessFit}
\title{Calculate goodness of fit}
\usage{
  goodnessFit(smpl, fit.pars, plot.fit = FALSE)
}
\arguments{
  \item{smpl}{sample vector}

  \item{fit.pars}{results from qFit()}

  \item{plot.fit}{Do you want to show fit plots?}
}
\description{
  Calculate goodness of fit
}

