\name{inv.trans.g}
\alias{inv.trans.g}
\title{Inverse of Tukey's g transformation (When h=0 and g is not 0)
When g=0 there is no need for an inverse since T(x)=x}
\usage{
  inv.trans.g(x, g)
}
\arguments{
  \item{x}{numeric vector}

  \item{g}{g parameter}
}
\description{
  Inverse of Tukey's g transformation (When h=0 and g is
  not 0) When g=0 there is no need for an inverse since
  T(x)=x
}

