\name{pgh}
\alias{pgh}
\title{Distribution function for Tukey's g-and-h
Currently working for h=0 only}
\usage{
  pgh(x, g = 0, A = 0, B = 1)
}
\arguments{
  \item{x}{numeric}

  \item{g}{parameter}

  \item{h}{parameter}

  \item{A}{parameter}

  \item{B}{parameter}
}
\description{
  Distribution function for Tukey's g-and-h Currently
  working for h=0 only
}

