\name{qgh}
\alias{qgh}
\title{Quantile function for Tukey's g-and-h
Currently working for h >= 0 only}
\usage{
  qgh(p, g = 0, h = 0, A = 0, B = 1)
}
\arguments{
  \item{p}{probs in (0, 1)}

  \item{g}{parameter}

  \item{h}{parameter}

  \item{A}{parameter}

  \item{B}{parameter}
}
\description{
  Quantile function for Tukey's g-and-h Currently working
  for h >= 0 only
}

