\name{quantile.table}
\alias{quantile.table}
\title{Summarize quantiles for a given sample}
\usage{
  quantile.table(smpl, max.depth)
}
\arguments{
  \item{smpl}{sample vector}

  \item{max.depth}{Use extremes?}
}
\description{
  Internal function Depends on tukey.quantiles() Depends on
  get.depth()
}

