\name{rgh}
\alias{rgh}
\title{Generate random numbers from Tukey's g-and-h}
\usage{
  rgh(n, g = 0, h = 0, A = 0, B = 1)
}
\arguments{
  \item{n}{n of samples to generate}

  \item{g}{parameter}

  \item{h}{parameter}

  \item{A}{parameter}

  \item{B}{parameter}
}
\description{
  Generate random numbers from Tukey's g-and-h
}

