% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGvsGH.R
\name{testGvsGH}
\alias{testGvsGH}
\title{Compute simulation-based \emph{p}-value of log-likelihood ratio test}
\usage{
testGvsGH(x, nsim, verbose = "vv")
}
\arguments{
\item{x}{data.}

\item{nsim}{number of Monte Carlo simulations}

\item{verbose}{function verbosity. Values \code{v}, \code{vv} and \code{vvv} are admitted,
whereas other values (such as \code{""} or \code{FALSE}) will make the function
silent.}
}
\value{
Object of class \code{testGvsGH}.
}
\description{
Compute simulation-based \emph{p}-value of log-likelihood ratio test
\insertCite{bee2021b}{tukeyGH}.
}
\examples{
\dontrun{

data(EPWS2014)
# Warning: the following code may take up to 30 mins to be run
testGvsGH(EPWS2014, 30)
}

}
\references{
\insertAllCited{}
}
