% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyGH-package.R
\docType{package}
\encoding{UTF-8}
\name{tukeyGH-package}
\alias{tukeyGH-package}
\alias{_PACKAGE}
\title{Tukey's g-and-h probability distribution}
\description{
It provides distribution, density and quantile functions of the
    Tukey's g-and-h probability distribution, as well as functions for random
    number generation, parameter estimation and testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-santi/tukeyGH}
  \item Report bugs at \url{https://github.com/f-santi/tukeyGH/issues}
}

}
\author{
\strong{Maintainer}: Flavio Santi \email{flavio.santi@univr.it} (\href{https://orcid.org/0000-0002-2014-1981}{ORCID})

Authors:
\itemize{
  \item Marco Bee \email{marco.bee@unitn.it} (\href{https://orcid.org/0000-0002-9579-3650}{ORCID})
  \item Julien Hambuckers \email{jhambuckers@uliege.be} (\href{https://orcid.org/0000-0003-4567-5477}{ORCID})
  \item Luca Trapin \email{luca.trapin@unibo.it} (\href{https://orcid.org/0000-0001-8613-0962}{ORCID})
}

}
